refactor(config): add `BaseConfig` class to enable multiple configurations
